Replace incoherent module with the new version.
